MS SQL Server was migrated
