--- asyncdb/providers/__init__.py
@@ -1,6 +1,0 @@
-"""
-AsyncDB Providers.
-"""
-from .base import InitProvider, BaseProvider
-
-__all__ = ['InitProvider', 'BaseProvider', ]
